+++ src/SmartHopper.Config/Models/TextChatModel.cs
@@ -56,28 +56,4 @@ namespace SmartHopper.Config.Models
         public string ToolCallId { get; set; } = "";
         public List<KeyValuePair<string, string>> ToolResult { get; set; }
     }
-
-    //public class ImageChatModel : IChatModel
-    //{
-    //    public bool Inbound { get; set; }
-    //    public bool Read { get; set; }
-    //    public DateTime Time { get; set; }
-    //    public string Author { get; set; }
-    //    public string Type { get; } = "image";
-
-    //    public Image Image { get; set; }
-    //    public string ImageName { get; set; }
-    //}
-
-    //public class AttachmentChatModel : IChatModel
-    //{
-    //    public bool Inbound { get; set; }
-    //    public bool Read { get; set; }
-    //    public DateTime Time { get; set; }
-    //    public string Author { get; set; }
-    //    public string Type { get; } = "attachment";
-
-    //    public byte[] Attachment { get; set; }
-    //    public string Filename { get; set; }
-    //}
 }